+++ Kernel/Arch/CPU.h
@@ -9,6 +9,13 @@
 
 #include <AK/Platform.h>
 
+#define PAGE_MASK (~(FlatPtr)0xfffu)
+
+#define LSW(x) ((u32)(x)&0xFFFF)
+#define MSW(x) (((u32)(x) >> 16) & 0xFFFF)
+#define LSB(x) ((x)&0xFF)
+#define MSB(x) (((x) >> 8) & 0xFF)
+
 #if ARCH(X86_64) || ARCH(I386)
 #    include <Kernel/Arch/x86/CPU.h>
 #elif ARCH(AARCH64)
