+++ Kernel/Arch/aarch64/CPU.h
@@ -6,8 +6,6 @@
 
 #pragma once
 
-#define PAGE_MASK (~(FlatPtr)0xfffu)
-
 namespace Kernel {
 
 void drop_to_exception_level_1();
