+++ Kernel/Arch/x86/CPU.h
@@ -18,7 +18,6 @@ VALIDATE_IS_X86()
 /* Map IRQ0-15 @ ISR 0x50-0x5F */
 #define IRQ_VECTOR_BASE 0x50
 #define GENERIC_INTERRUPT_HANDLERS_COUNT (256 - IRQ_VECTOR_BASE)
-#define PAGE_MASK (~(FlatPtr)0xfffu)
 
 namespace Kernel {
 
@@ -35,11 +34,6 @@ inline u32 get_iopl_from_eflags(u32 eflags)
 
 DescriptorTablePointer const& get_gdtr();
 DescriptorTablePointer const& get_idtr();
-
-#define LSW(x) ((u32)(x)&0xFFFF)
-#define MSW(x) (((u32)(x) >> 16) & 0xFFFF)
-#define LSB(x) ((x)&0xFF)
-#define MSB(x) (((x) >> 8) & 0xFF)
 
 constexpr FlatPtr page_base_of(FlatPtr address)
 {
